requirements.txt
absl-py==2.4.0
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiomultiprocess==0.9.1
aiosignal==1.4.0
annotated-doc==0.0.4
anyio==4.12.1
async-timeout==5.0.1
attrs==25.4.0
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
cuda-bindings==12.9.4
cuda-pathfinder==1.2.2
datasets==4.5.0
dill==0.4.0
exceptiongroup==1.3.1
faiss-cpu==1.13.2
filelock==3.20.0
frozenlist==1.8.0
fsspec==2025.10.0
grpcio==1.78.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.4.1
idna==3.11
Jinja2==3.1.6
joblib==1.5.3
Markdown==3.10.2
markdown-it-py==4.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mpmath==1.3.0
multidict==6.7.1
multiprocess==0.70.18
networkx==3.4.2
numpy==1.26.4
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.4.5
nvidia-nvtx-cu12==12.8.90
packaging @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_packaging_1769093650/work
pandas==2.3.3
pillow==12.0.0
propcache==0.4.1
protobuf==6.33.5
psutil==7.2.2
pyarrow==23.0.1
Pygments==2.19.2
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
regex==2026.1.15
requests==2.32.5
rich==14.3.2
safetensors==0.7.0
scikit-learn==1.7.2
scipy==1.15.3
shellingham==1.5.4
six==1.17.0
sympy==1.14.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
threadpoolctl==3.6.0
tokenizers==0.22.2
torch==2.10.0+cu128
torchvision==0.25.0+cu128
tqdm==4.67.3
transformers==5.2.0
triton==3.6.0
typer==0.24.0
typer-slim==0.24.0
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
Werkzeug==3.1.5
xxhash==3.6.0
yarl==1.22.0
